@@ -2,39 +2,35 @@
   "name": "awesome-kotlin",
   "version": "1.0.0",
   "description": "Awesome list for Kotlin",
-  "main": "app/src/main.tsx",
+  "main": "app/src/index.tsx",
   "dependencies": {
     "copy-to-clipboard": "^3.3.3",
-    "core-js": "^3.38.1",
-    "normalize.css": "^8.0.1",
+    "immer": "^10.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router": "^6.26.1",
-    "react-router-dom": "^6.26.1"
+    "zustand": "^4.5.5"
   },
   "devDependencies": {
     "@types/node": "^22.5.4",
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
-    "autoprefixer": "^10.4.20",
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
+    "css-minimizer-webpack-plugin": "^7.0.0",
     "html-webpack-plugin": "^5.6.0",
-    "less": "^4.2.0",
-    "less-loader": "^12.2.0",
-    "postcss": "^8.4.45",
-    "postcss-loader": "^8.1.1",
+    "mini-css-extract-plugin": "^2.9.1",
+    "sass": "^1.78.0",
+    "sass-loader": "^16.0.1",
     "style-loader": "^4.0.0",
     "ts-loader": "^9.5.1",
-    "typescript": "^5.5.4",
+    "typescript": "^5.6.2",
     "webpack": "^5.94.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.1.0"
+    "webpack-cli": "^5.1.4"
   },
   "scripts": {
     "prestart": "yarn install",
     "prebuild": "yarn install",
-    "start": "webpack serve --mode development --env NODE_ENV=development --env SOURCE_MAP=eval-source-map",
+    "start": "webpack watch --mode development --env NODE_ENV=development --env SOURCE_MAP=eval-source-map",
     "build": "webpack --mode production --env NODE_ENV=production --env SOURCE_MAP=source-map"
   },
   "repository": {
@@ -44,11 +40,11 @@
   "keywords": [
     "kotlin"
   ],
-  "author": "Ibragimov Ruslan, Labushkina Hanna",
+  "author": "Ibrahimau Ruslan",
   "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/Heapy/awesome-kotlin/issues"
   },
   "homepage": "https://github.com/Heapy/awesome-kotlin#readme",
-  "packageManager": "yarn@4.3.1"
+  "packageManager": "yarn@4.4.1"
 }
