@@ -1,17 +1,26 @@
 package usecases.version
 
-import HttpClientModule
-import XmlModule
-import di.bean
+import infra.HttpClientModule
+import io.heapy.komok.tech.di.lib.Module
+import infra.serialization.XmlModule
 
+@Module
 open class KotlinVersionModule(
     private val xmlModule: XmlModule,
     private val httpClientModule: HttpClientModule,
 ) {
-    open val versionFetcher by bean<KotlinVersionFetcher> {
-        MavenCentralKotlinVersionFetcher(
-            xmlMapper = xmlModule.xmlMapper.get,
-            httpClient = httpClientModule.httpClient.get,
+    open val kotlinVersionFetcher: KotlinVersionFetcher by lazy {
+        CachedVersionFetcher(
+            kotlinVersionFetcher = MavenCentralKotlinVersionFetcher(
+                xml = xmlModule.xml,
+                httpClient = httpClientModule.httpClient,
+            )
+        )
+    }
+
+    open val route by lazy {
+        KotlinVersionRoute(
+            kotlinVersionFetcher = kotlinVersionFetcher,
         )
     }
 }
