@@ -1,9 +1,8 @@
 package usecases.kug
 
-import io.ktor.server.application.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import ktor.KtorRoute
+import infra.ktor.KtorRoute
 
 class GetKugsRoute(
     private val kugDao: KugDao,
