@@ -1,22 +1,22 @@
 package usecases.links
 
-import di.bean
-import io.ktor.server.application.call
+import io.heapy.komok.tech.di.lib.Module
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Routing
 import io.ktor.server.routing.get
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
-import ktor.KtorRoute
+import infra.ktor.KtorRoute
 
+@Module
 open class LinksModule {
-    open val linkSource by bean {
+    open val linkSource by lazy {
         LinkSource().get()
     }
 
-    open val route by bean {
+    open val route by lazy {
         LinksRoute(
-            links = linkSource.get,
+            links = linkSource,
         )
     }
 }
@@ -25,69 +25,87 @@ class LinksRoute(
     private val links: List<CategoryV1>,
 ) : KtorRoute {
     override fun Routing.install() {
-        get("/links") {
-            val isAwesome = call.request.queryParameters["awesome"].toBoolean()
+        get("/api/links") {
+            val isAwesome = call.request.queryParameters["awesome"]?.toBoolean() ?: false
+            val isKugs = call.request.queryParameters["kugs"]?.toBoolean() ?: false
+            val query = call.request.queryParameters["query"]
 
-            val allLinks: List<CategoryDto> = if (isAwesome) {
-                links.map { cat ->
-                    val newSub = cat.subcategories
-                        .map { sub ->
-                            sub.copy(links = sub.links.filter { link -> link.awesome }.toMutableList())
-                        }
-                        .filter { sub ->
-                            sub.links.isNotEmpty()
-                        }
-                        .toMutableList()
-
-                    cat.copy(subcategories = newSub)
-                }.filter { it.subcategories.isNotEmpty() }
-                    .map { category -> category.toDto() }
+            val filteredLinks = if (isAwesome) {
+                if (query != null) {
+                    links.filterByQuery(query)
+                } else {
+                    links.filterByAwesome()
+                }
+            } else if (isKugs) {
+                links.filterByKug().filterByQuery(query)
             } else {
-                links.map { category -> category.toDto() }
+                links.filterByQuery(query)
             }
 
-            call.respond(allLinks)
+            call.respond(filteredLinks)
         }
     }
 }
 
-class LinkSource {
-    private val files = listOf(
-        "Links.json",
-        "Libraries.json",
-        "Projects.json",
-        "Android.json",
-        "JavaScript.json",
-        "Native.json",
-        "UserGroups.json"
-    )
+fun List<CategoryV1>.filterByAwesome(): List<CategoryV1> {
+    return this.mapNotNull { category ->
+        val subcategories = category.subcategories.mapNotNull { subcategory ->
+            val links = subcategory.links.filter { linkV1 ->
+                linkV1.state == LinkStateV1.AWESOME
+            }
 
-    fun get(): List<CategoryV1> {
-        return files.map { file ->
-            val json = LinkSource::class.java
-                .classLoader
-                .getResource("links/$file")
-                .readText()
-            Json.decodeFromString(json)
+            if (links.isNotEmpty()) subcategory.copy(links = links) else null
         }
+
+        if (subcategories.isNotEmpty()) category.copy(subcategories = subcategories) else null
+    }
+}
+fun List<CategoryV1>.filterByKug(): List<CategoryV1> {
+    return this.filter { categoryV1 ->
+        categoryV1.name == "Kotlin User Groups"
+    }
+}
+
+fun List<CategoryV1>.filterByQuery(query: String?): List<CategoryV1> {
+    return if (query != null) {
+        val searchTerm = query.lowercase()
+
+        this.mapNotNull { category ->
+            val subcategories = category.subcategories.mapNotNull { subcategory ->
+                val links = subcategory.links.filter { link ->
+                    link.name.contains(searchTerm, ignoreCase = true) ||
+                        link.desc?.contains(searchTerm, ignoreCase = true) == true ||
+                        link.tags.any { tag -> tag.contains(searchTerm, ignoreCase = true) }
+                }
+                if (links.isNotEmpty()) subcategory.copy(links = links) else null
+            }
+            if (subcategories.isNotEmpty()) category.copy(subcategories = subcategories) else null
+        }
+    } else {
+        this
+    }
+}
+
+class LinkSource {
+    fun get(): List<CategoryV1> {
+        val json = LinkSource::class.java
+            .classLoader
+            .getResource("links/links.json")
+            .readText()
+        return Json.decodeFromString(json)
     }
 }
 
 @Serializable
 data class LinkV1(
-    val name: String? = null,
-    val github: String? = null,
-    val bitbucket: String? = null,
-    val kug: String? = null,
+    val name: String,
     val href: String? = null,
     val desc: String? = null,
     val platforms: List<PlatformTypeV1> = emptyList(),
     val tags: List<String> = emptyList(),
     val star: Int? = null,
     val update: String? = null,
-    val archived: Boolean = false,
-    val unsupported: Boolean = false,
-    val awesome: Boolean = false
+    val state: LinkStateV1,
 )
 
 @Serializable
@@ -102,75 +120,21 @@ enum class PlatformTypeV1 {
 }
 
 @Serializable
+enum class LinkStateV1 {
+    AWESOME,
+    UNSUPPORTED,
+    ARCHIVED,
+    DEFAULT,
+}
+
+@Serializable
 data class SubcategoryV1(
     val name: String,
-    val links: MutableList<LinkV1>
+    val links: List<LinkV1>
 )
 
 @Serializable
 data class CategoryV1(
     val name: String,
-    val subcategories: MutableList<SubcategoryV1>
+    val subcategories: List<SubcategoryV1>
 )
-
-data class LinkDto(
-    val name: String,
-    val href: String,
-    val desc: String,
-    val platforms: List<PlatformTypeV1>,
-    val tags: Set<String>,
-    val star: Int? = null,
-    val update: String? = null,
-    val state: LinkStateDto
-)
-
-enum class LinkStateDto {
-    AWESOME,
-    UNSUPPORTED,
-    ARCHIVED,
-    DEFAULT
-}
-
-private fun LinkV1.toDto(): LinkDto {
-    val state = when {
-        awesome -> LinkStateDto.AWESOME
-        archived -> LinkStateDto.ARCHIVED
-        unsupported -> LinkStateDto.UNSUPPORTED
-        else -> LinkStateDto.DEFAULT
-    }
-
-    return LinkDto(
-        name = name ?: error("Link should have a name [$this]"),
-        href = href ?: error("Link should have a href [$this]"),
-        desc = desc ?: "",
-        platforms = platforms,
-        tags = tags.toSet(),
-        star = star,
-        update = update,
-        state = state
-    )
-}
-
-data class SubcategoryDto(
-    val name: String,
-    val links: List<LinkDto>
-)
-
-data class CategoryDto(
-    val name: String,
-    val subcategories: List<SubcategoryDto>
-)
-
-fun CategoryV1.toDto(): CategoryDto {
-    return CategoryDto(
-        name = name,
-        subcategories = subcategories.map { it.toDto() }
-    )
-}
-
-private fun SubcategoryV1.toDto(): SubcategoryDto {
-    return SubcategoryDto(
-        name = name,
-        links = links.map { it.toDto() }
-    )
-}
