@@ -1,11 +1,10 @@
 package usecases.signup
 
-import JooqModule
+import infra.db.JooqModule
 import at.favre.lib.crypto.bcrypt.BCrypt
 import at.favre.lib.crypto.bcrypt.LongPasswordStrategies
-import di.bean
+import io.heapy.komok.tech.di.lib.Module
 import io.ktor.http.HttpStatusCode
-import io.ktor.server.application.call
 import io.ktor.server.request.receive
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Routing
@@ -13,13 +12,14 @@ import io.ktor.server.routing.post
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
-import ktor.KtorRoute
+import infra.ktor.KtorRoute
 import java.security.SecureRandom
 
+@Module
 open class RegisterModule(
     private val jooqModule: JooqModule,
 ) {
-    open val bcryptHasher by bean<BCrypt.Hasher> {
+    open val bcryptHasher by lazy<BCrypt.Hasher> {
         BCrypt.with(
             BCrypt.Version.VERSION_2A,
             SecureRandom(),
@@ -27,14 +27,14 @@ open class RegisterModule(
         )
     }
 
-    open val kotlinerDao by bean<KotlinerDao> {
-        DefaultKotlinerDao(jooqModule.dslContext.get)
+    open val kotlinerDao by lazy<KotlinerDao> {
+        DefaultKotlinerDao(jooqModule.dslContext)
     }
 
-    open val route by bean {
+    open val route by lazy {
         RegisterRoute(
-            bcryptHasher = bcryptHasher.get,
-            kotlinerDao = kotlinerDao.get,
+            bcryptHasher = bcryptHasher,
+            kotlinerDao = kotlinerDao,
         )
     }
 }
