@@ -1,8 +1,17 @@
 import * as React from "react";
-import {KotlinVersion} from "../version/KotlinVersion";
-import "./bar.less";
+import "./bar.scss";
+import {useVersionsStore} from "./useVersionsStore";
+import {memo, useEffect} from "react";
+import KotlinVersion from "../version/kotlin_version";
 
-export function Bar({versions}: Props) {
+function Bar() {
+  const fetchVersions = useVersionsStore(state => state.fetchVersions);
+  const versions = useVersionsStore(state => state.versions);
+
+  useEffect(() => {
+    fetchVersions();
+  }, [fetchVersions]);
+
   return (
     <section className="bar">
       <div className="bar_wrapper">
@@ -12,6 +21,4 @@ export function Bar({versions}: Props) {
   )
 }
 
-export interface Props {
-  readonly versions: string[];
-}
+export default memo(Bar);
