@@ -1,7 +1,9 @@
 import * as React from "react";
-import {LinksPageComponent} from "../../components/page_wrapper/page_wrapper";
-import {all} from "../../links";
+import LinksPage, {PageType} from "../../components/page_wrapper/page_wrapper";
+import {memo} from "react";
 
-export default function Resources() {
-  return <LinksPageComponent displayLinks={all} searchLinks={all}/>
+function Resources() {
+  return <LinksPage pageType={PageType.ALL}/>
 }
+
+export default memo(Resources);
