@@ -2,6 +2,8 @@ const path = require("path");
 const webpack = require("webpack");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const CssMinimizerPlugin = require("css-minimizer-webpack-plugin");
 
 module.exports = function (options = {}) {
   // Settings
@@ -11,89 +13,89 @@ module.exports = function (options = {}) {
 
   const APP_DIR = path.resolve(__dirname, "app");
 
-  console.error(`
-Build started with following configuration:
-===========================================
-→ NODE_ENV: ${NODE_ENV}
-→ SOURCE_MAP: ${SOURCE_MAP}
-`);
-
   return {
     entry: {
       app: [
-        path.resolve(APP_DIR, "src", "main.tsx")
-      ]
+        path.resolve(APP_DIR, "src", "index.tsx"),
+      ],
     },
     output: {
       path: path.resolve(__dirname, "dist"),
       filename: "[name].js?[contenthash]",
       chunkFilename: "[name].bundle.js?[chunkhash]",
-      publicPath: "/"
+      publicPath: "/",
+      clean: true,
     },
     optimization: {
+      minimizer: [
+        "...",
+        new CssMinimizerPlugin(),
+      ],
       splitChunks: {
-        chunks: "all"
-      }
+        chunks: "all",
+      },
     },
     resolve: {
-      extensions: [".ts", ".tsx", ".js"]
+      extensions: [".ts", ".tsx", ".js"],
     },
     bail: false,
     devtool: SOURCE_MAP,
     module: {
-      rules: [{
-        test: /\.tsx?$/,
-        use: [{
-          loader: "ts-loader"
-        }]
-      }, {
-        test: /\.less$/,
-        use: [{
-          loader: "style-loader"
-        }, {
-          loader: "css-loader"
-        }, {
-          loader: "postcss-loader",
-          options: {
-            postcssOptions: {
-              plugins: [["autoprefixer"]]
+      rules: [
+        {
+          test: /\.tsx?$/,
+          use: [
+            {
+              loader: "ts-loader",
             },
-          }
-        }, {
-          loader: "less-loader"
-        }]
-      }, {
-        test: /\.(png|jpg|gif|svg)$/,
-        type: "asset",
-        parser: {
-          dataUrlCondition: {
-            maxSize: 32768
-          }
+          ],
+        },
+        {
+          test: /\.scss$/,
+          use: [
+            MiniCssExtractPlugin.loader,
+            {
+              loader: "css-loader",
+            },
+            {
+              loader: "sass-loader",
+            },
+          ],
+        },
+        {
+          test: /\.(png|jpg|gif|svg)$/,
+          type: "asset",
+          parser: {
+            dataUrlCondition: {
+              maxSize: 1024,
+            },
+          },
         }
-      }]
+      ]
     },
     plugins: createListOfPlugins({NODE_ENV}, APP_DIR),
-    devServer: {
-      historyApiFallback: true
-    }
   }
 };
 
 function createListOfPlugins({NODE_ENV}, APP_DIR) {
   return [
+    new MiniCssExtractPlugin({
+      filename: '[name]_[contenthash].css',
+      chunkFilename: '[id]_[contenthash].css',
+    }),
     new HtmlWebpackPlugin({
       template: path.resolve(APP_DIR, "index.html"),
-      hash: true
+      hash: true,
     }),
     new CopyWebpackPlugin({
       patterns: [
-        {from: path.resolve(APP_DIR, "icons")}
-      ]
+        {from: path.resolve(APP_DIR, "icons")},
+      ],
     }),
     new webpack.DefinePlugin({
       "process.env": {
-        "NODE_ENV": JSON.stringify(NODE_ENV)
-      }
-    })
+        "NODE_ENV": JSON.stringify(NODE_ENV),
+      },
+    }),
   ];
 }
