@@ -15,11 +15,9 @@ import java.util.Date
 
 interface RssGenerator {
     fun generate(articles: List<Article>, name: String): String
-
-    companion object
 }
 
-private class DefaultRssGenerator : RssGenerator {
+class DefaultRssGenerator : RssGenerator {
     override fun generate(articles: List<Article>, name: String): String {
         val feed = SyndFeedImpl().apply {
             title = "Kotlin Programming Language"
@@ -65,8 +63,3 @@ private fun toSyndEntry(article: Article): SyndEntry {
         }
     }
 }
-
-fun RssGenerator.Companion.default(): RssGenerator {
-    return DefaultRssGenerator()
-}
-
