@@ -2,15 +2,12 @@ package link.kotlin.scripts
 
 import link.kotlin.scripts.dsl.Article
 import link.kotlin.scripts.model.ApplicationConfiguration
-import link.kotlin.scripts.model.default
 
 interface SitemapGenerator {
     fun generate(articles: List<Article>): String
-
-    companion object
 }
 
-private class DefaultSitemapGenerator(
+class DefaultSitemapGenerator(
     private val configuration: ApplicationConfiguration
 ) : SitemapGenerator {
 
@@ -23,12 +20,4 @@ private class DefaultSitemapGenerator(
             }
         }
     }
-}
-
-fun SitemapGenerator.Companion.default(
-    configuration: ApplicationConfiguration = ApplicationConfiguration.default()
-): SitemapGenerator {
-    return DefaultSitemapGenerator(
-        configuration = configuration
-    )
 }
