@@ -3,9 +3,9 @@ package link.kotlin.scripts.import
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import kotlinx.coroutines.runBlocking
+import link.kotlin.scripts.ApplicationFactory
 import link.kotlin.scripts.utils.HttpClient
 import link.kotlin.scripts.utils.body
-import link.kotlin.scripts.utils.default
 import link.kotlin.scripts.utils.parseInstant
 import org.apache.http.client.methods.HttpGet
 import org.apache.http.client.utils.URIBuilder
@@ -29,9 +29,8 @@ class Readability(
 
 fun main() = runBlocking {
     val mapper = jacksonObjectMapper()
-    val readability = Readability(
-        HttpClient.default()
-    )
+    val httpClient = ApplicationFactory().httpClient
+    val readability = Readability(httpClient)
 
     val response = readability.getArticle("https://blog.jetbrains.com/kotlin/2017/03/kotlin-1-1-1-is-out/")
     val res = mapper.readValue<ReadabilityResponse>(response)
